refactor: extract classifyRacer to certificate-stats, consolidate den constants
Move CertTier type and classifyRacer() from CertificateView into
certificate-stats.ts where they can be unit tested independently.
Move DEN_SINGULAR and DEN_ACCENT to den-utils.ts alongside DEN_IMAGES.
